requirements.txt
python-dotenv
crewai
streamlit
langchain_openai
crewai_tools
langchain_community

    